% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_doc.R
\docType{data}
\name{adm_boundaries}
\alias{adm_boundaries}
\title{Example SpatialPolygonsDataFrame of country boundaries}
\format{
SpatialPolygonsDataFrame with 9 features.
\describe{
\item{features}{SpatialPolygons, 9.}
\item{data.frame}{9 rows, 11 columns.}
}
}
\usage{
adm_boundaries
}
\description{
A SpatialPolygonsDataFrame of 9 countries from South America.
}
\examples{
data("adm_boundaries", package = "rangemap")
adm_boundaries
}
\keyword{datasets}
