% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{aoo}
\alias{aoo}
\title{Area of occupancy of a species as defined by the IUCN}
\usage{
aoo(occ_pr, species)
}
\arguments{
\item{occ_pr}{SpatialPointsDataFrame of ocurrence records. Projection must be
one that allows safe calculation of areas (e.g., Eckert IV; EPSG:54012)}

\item{species}{(character) scientific name of the species.}
}
\value{
A list comtaining a SpatialPolygonsDataFrame of the area of occupancy, and
a vector with the areas in km2 of the spatial polygons resulted.
}
\description{
Area of occupancy of a species as defined by the IUCN
}
\examples{
# data
data("occ_p", package = "rangemap")
occ <- unique(occ_p)
WGS84 <- sp::CRS("+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
occ_sp <- sp::SpatialPointsDataFrame(coords = occ[, 2:3], data = occ,
                                     proj4string = WGS84)

ECK4 <- sp::CRS("+proj=eck4 +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs")
occ_pr <- sp::spTransform(occ_sp, ECK4)

sp <- as.character(occ[1, 1])

# AOO
aoo_pe <- aoo(occ_pr = occ_pr, species = sp)
}
