% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_sdm.R
\name{construct_sdm}
\alias{construct_sdm}
\title{Using Maxent to construct Species Distribution Model (SDM)}
\usage{
construct_sdm(
  species_name,
  data_dir = "data_processed/species_data",
  n_cl = 10,
  r = 20,
  model_dir = "models",
  env,
  maxent_args,
  result_dir = "results/gis/prediction_avg",
  ...
)
}
\arguments{
\item{species_name}{The name of the species using in data file, also for SDM folder.}

\item{data_dir}{The path to the directory containing species occurrence data in .geojson format.}

\item{n_cl}{The number of clusters for parallel processing (default is 10).}

\item{model_dir}{The path to the directory store species model, default is "models".}

\item{result_dir}{The path to the directory to store average predicted raster file, default is "results/gis/prediction_avg".}
}
\value{
The function constructs the SDM and does not return any value. It saves
prediction results in the "results/gis/prediction_avg" directory as default.
}
\description{
This function using MaxEnt to constructs a Species Distribution Model (SDM) for a given species
based on species occurrence data in a .geojson file. It performs MaxEnt modeling
and exports prediction results. The SDM is built for multiple replicates (default r is 20),
and prediction averages are calculated.
}
\examples{
\dontrun{
# Example usage:
construct_sdm("Example_Species", path = "data_processed/species_data", n_cl = 5)
}

}
