% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_codes.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{List of country names and ISO codes}
\format{
A data frame with 247 rows and 4 columns.
\describe{
\item{Country_or_Area_Name}{character, names.}
\item{ISO_ALPHA.2_Code}{character, country code.}
\item{ISO_ALPHA.3_Code}{character, country code.}
\item{ISO_Numeric_Code_UN_M49_Numerical_Code}{numeric, country numeric codes.}
}
}
\source{
\url{https://www.nationsonline.org/oneworld/country_code_list.htm}
}
\usage{
country_codes
}
\description{
A dataset containing codes for identifying countries according
to ISO norms.
}
\examples{
data("country_codes", package = "rangemap")
head(country_codes)
}
\keyword{datasets}
