% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_doc.R
\docType{data}
\name{cvehull_range}
\alias{cvehull_range}
\title{Example of sp_range* object based on concave hulls}
\format{
sp_range_iucn with 6 slots.
\describe{
\item{name}{character, name identifying the origin of the object.}
\item{summary}{data.frame, summary of results.}
\item{species_range}{SpatialPolygonsDataFrame of species range.}
\item{species_unique_records}{SpatialPointsDataFrame of species occurrences.}
\item{extent_of_occurrence}{SpatialPolygonsDataFrame of species extent of occurrence.}
\item{area_of_occupancy}{SpatialPolygonsDataFrame of species area of occupancy.}
}
}
\usage{
cvehull_range
}
\description{
A sp_range_iucn object containing the results of the function
\code{\link{rangemap_hull}}.
}
\examples{
data("cvehull_range", package = "rangemap")
summary(cvehull_range)
}
\keyword{datasets}
