% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{env}
\alias{env}
\title{Data of environment variables}
\format{
A list of stacked raster.
}
\source{
Environment variables download from "A multi-temporal and terrestrial environmental dataset of Taiwan". Includ Bio01, Bio02, Bio04, Bio12, Bio15, Bio16, Aspect, ASR, ELE, ELESD, Slope, BL, BU, FF, FO, FW, MD, UB, WB, WL, DFW.
}
\usage{
env
}
\description{
This data contains 21 environment variables used in MaxEnt.
}
\keyword{datasets}
