% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{eoo}
\alias{eoo}
\title{Extent of occurrence of a species based on convex hull polygons}
\usage{
eoo(occurrences, polygons)
}
\arguments{
\item{occurrences}{a data.frame containing geographic coordinates of species
occurrences, columns must be: Species, Longitude, and Latitude. Geographic
coordinates must be in decimal degrees.}

\item{polygons}{SpatialPolygon object to clip convex hulls to these limits.
Projection of this object defines projection of the extent of occurrence. This
projection must be one that allows safe calculation of areas (e.g., Eckert IV;
EPSG:54012)}
}
\value{
A list comtaining a SpatialPolygonsDataFrame of the extent of occurrence, and
a vector with the areas in km2 of the spatial polygons resulted.
}
\description{
Extent of occurrence of a species based on convex hull polygons
}
\examples{
# occurrences
data("occ_p", package = "rangemap")
occ <- unique(occ_p)

# polygons
data("wrld_simpl", package = "maptools")
ECK4 <- sp::CRS("+proj=eck4 +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs")
poly_pr <- sp::spTransform(wrld_simpl, ECK4)

# EOO
eoo_pe <- eoo(occurrences = occ, polygons = poly_pr)
}
