% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GADM_spoly}
\alias{GADM_spoly}
\title{Get SpatialPolygons of countries at distinct administrative levels}
\usage{
GADM_spoly(country_code, boundary_level, keep_data = FALSE)
}
\arguments{
\item{country_code}{(character) code of country or countries of interest.}

\item{boundary_level}{(numeric) level of administrative division to be considered.}

\item{keep_data}{(logical) whether or not to keep downloaded files. Default = \code{FALSE}.}
}
\value{
A SpatialPolygonsDataFrame from the GADM database at the level selected.
If an error occurs when downloading any of the spatial objects based on
\code{country_code}, the result is NULL.
}
\description{
Get SpatialPolygons of countries at distinct administrative levels
}
\examples{
map_gadm <- GADM_spoly(country_code = "UY", boundary_level = 0)
}
