% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_range_map_layer.R
\name{generate_range_map_layer}
\alias{generate_range_map_layer}
\title{Generate Range Map in .geojson Format Using Thresholds Select by Expert}
\usage{
generate_range_map_layer(
  species_list,
  r = 20,
  threshold_table,
  model_dir = "models",
  predict_dir = "results/gis/prediction_avg",
  output_dir_gis = "results/gis/range_map_ori",
  output_dir_png = NULL,
  ...
)
}
\arguments{
\item{species_list}{A character vector of species names for which range maps are generated.}

\item{r}{The number of replicates used in MaxEnt modeling (default is 20).}

\item{threshold_table}{A data.table containing expert-selected thresholds for each species.}

\item{model_dir}{The directory containing MaxEnt modeling results (default is "models").}

\item{predict_dir}{The directory containing prediction averages (default is "results/gis/prediction_avg").}

\item{output_dir_gis}{The directory where the range maps in .geojson format will be saved
(default is "results/gis/range_map_ori").}

\item{output_dir_png}{The directory where the range maps in .png format will be saved
(default is NULL).}
}
\value{
This function generates range maps and does not return any value.
}
\description{
This function generates species range map file in .geojson format for multiple species
using thresholds provided by an expert. The thresholds are based on MaxEnt modeling
results. The resulting range maps are saved in the specified output directory.
}
\examples{
\dontrun{
# Example usage:
species_list <- c("Species_A", "Species_B")
threshold_table <- fread("path/to/threshold_table.csv") \%>\% select(species, threshold_maxent)
generate_range_map_layer(species_list, r = 20, threshold_table)
}

}
