% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_species_data.R
\name{generate_species_data}
\alias{generate_species_data}
\title{Generate Species Data and Export .geojson Files}
\usage{
generate_species_data(
  sp_data,
  species_name,
  coord = c("longitude", "latitude"),
  crs_sp = 4326,
  path = "data_processed/species_data",
  ...
)
}
\arguments{
\item{sp_data}{Species occurrence data in data.frame or data.table format.}

\item{species_name}{Name of the species to be used in the file name.}

\item{coord}{A vector containing the column names of coordinates in the data.
The default is c("longitude", "latitude").}

\item{path}{the folder path store the .geojson file.}

\item{crs}{The EPSG code of coordinate reference system (CRS) of the data, specified as
a numeric vector. The default is 4326.}
}
\value{
This function generates .geojson files for each species and does not
return any value.
}
\description{
This function takes species occurrence data in data.frame or data.table format
and exports a .geojson file for each species. The .geojson files are saved in
the "species_data" directory under the "data_processed" folder as defult.
}
\examples{
\dontrun{
# Example usage:
data <- data.frame(
  longitude = c(120.27, 120.15, 121.19),
  latitude = c(23.22, 24.10, 23.59),
  other_column = c("A", "B", "C")
)
generate_species_data(sp_data = data, species_name = "Species_A")
}

}
