% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_species_range_map.R
\name{generate_species_range_map}
\alias{generate_species_range_map}
\title{Generate Species Range Map Plot}
\usage{
generate_species_range_map(range_map, species_name)
}
\arguments{
\item{range_map}{The spatial data frame for the species range map.}

\item{species_name}{The name of the species.}
}
\value{
A ggplot object representing the species range map.
}
\description{
This function generates a ggplot for a species range map using specified style.
}
\examples{
\dontrun{
sp <- "Species_A"
sp_range_map <- st_read("path/to/species_range_map.geojson")
generate_species_range_map(species_name = sp, range_map = sp_range_map)
ggsave(sprintf("\%s.png", sp),
       path = "results/plots/range_map_modified",
       width = 7, height = 9, dpi = 300)
}

}
