% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workflow.R
\name{get_workflow}
\alias{get_workflow}
\title{Save the example workflow R code to another location}
\usage{
get_workflow(file_path = "scripts", file_name = "workflow.R")
}
\arguments{
\item{file_path}{The destination path where you want to copy the example code. The default is "scripts"}

\item{file_name}{The file name of script saved. The default is "workflow.R"}
}
\description{
This function copies the example workflow R code from the package to a specified destination path.
}
\examples{
\dontrun{
# Example usage:
get_workflow(file_path = "scripts")
}

}
