% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{grid_1km}
\alias{grid_1km}
\title{Data of grid system}
\format{
A sf with 1km grid system.
}
\source{
1 km grid system of Taiwan, with GRIDID.x for 1 km grid ID and GRIDID.y for 5 km ID.
}
\usage{
grid_1km
}
\description{
This data contains the grid system used to subsample occurrence data.
}
\keyword{datasets}
