% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_espace.R
\name{isnested}
\alias{isnested}
\title{Helper function to test if lists are nested.}
\usage{
isnested(l)
}
\arguments{
\item{l}{list to be tested.}
}
\description{
Helper function to test if lists are nested.
}
