% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{LAEA_projection}
\alias{LAEA_projection}
\title{Prepare Lambert Azimuthal Equal-Area Projection}
\usage{
LAEA_projection(occ = NULL)
}
\arguments{
\item{occ}{matrix or SpatialPoint object conaining coordinates to serve as
a reference for the center of the projection. If not defined projection is
centered in 0, 0 for longitude and latitude. If a matrix is used columns must
"longitude" and "latitude", in that order.}
}
\value{
An object of class CRS.
}
\description{
Prepare Lambert Azimuthal Equal-Area Projection
}
\examples{
LAEA_projection()

data("occ_p", package = "rangemap")
occ <- unique(occ_p)[, 2:3]
LAEA_projection(occ)
}
