% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_required_data.R
\name{load_required_data}
\alias{load_required_data}
\title{Load required data for the RangeMap package}
\usage{
load_required_data()
}
\description{
This function loads essential data files required for the RangeMap package.
}
\examples{
\dontrun{
# Load data files when loading the RangeMap package
load_required_data()
}

}
