% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{maxent_args}
\alias{maxent_args}
\title{Data of maxent argument setting}
\format{
A vector of characters.
}
\source{
MaxEnt argumet used.
}
\usage{
maxent_args
}
\description{
This data contains maxent argument used to build SDM
}
\keyword{datasets}
