% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2binary_tb.R
\name{model2binary_tb}
\alias{model2binary_tb}
\title{Convert Continuous Predicted Values to Binary Based on MaxEnt Thresholds}
\usage{
model2binary_tb(
  model_dir = "models",
  predicted_dir = "results/gis/prediction_avg",
  species_name
)
}
\arguments{
\item{model_dir}{The directory containing MaxEnt modeling results (default is "models").}

\item{predicted_dir}{The directory containing continuous predicted values (default is "results/gis/prediction_avg").}

\item{species_name}{The species identifier used for the model folder and predicted values. It serves as the unique name associated with the species throughout the modeling and mapping process.}
}
\value{
A data.frame with binary values for each threshold and coordinates.
}
\description{
This function takes the continuous predicted values generated by MaxEnt for a species,
applies 11 threshold values provided by MaxEnt, and returns a data.frame with binary
values for each threshold along with the coordinates.
}
\examples{
\dontrun{
# Example usage:
model2binary_tb(model_dir = "models", predicted_dir = "results/gis/prediction_avg", species_name = "Species_A")
}

}
