% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_fig.R
\name{north_arrow}
\alias{north_arrow}
\title{North arrow for map plots}
\usage{
north_arrow(position = "topright", xlim, ylim, Ncex = 0.6)
}
\arguments{
\item{position}{(character or numeric) position of the North arrow. If character, options
are: "topright", "topleft", "bottomleft", or "bottomright". Default = "topright".}

\item{xlim}{(numeric) vector of two numbers indicating the x limits of the plotting area.}

\item{ylim}{(numeric) vector of two numbers indicating the y limits of the plotting area.}

\item{Ncex}{(numeric) cex for the North label (N).}

\item{exproting}{(logical) whether or not the map will be exported as a figure.}
}
\description{
north_arrow plots a North arrow in user defined places in a map.
}
