% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ2gridCentroid.R
\name{occ2gridCentroid}
\alias{occ2gridCentroid}
\title{Turn coordinates of species occurrence to grid centroid, keeping all columns of species occurrence data}
\usage{
occ2gridCentroid(
  occ_data,
  coords_arg = c("X", "Y"),
  occ_crs = 4326,
  grid_sf,
  grid_id_x = "GRIDID.x",
  grid_id_y = "GRIDID.y",
  grid_crs = 3826
)
}
\arguments{
\item{occ_data}{A data.table of species occurrence data with coordinates.}

\item{coords_arg}{A vector of variable names of coordinates used in occ_data. Default is c("X", "Y").}

\item{occ_crs}{EPSG code of occ_data. Default is 4326.}

\item{grid_sf}{An sf object of small size grid layer.}

\item{grid_id_x}{Small size grid identifier used in grid_sf. Default is "GRIDID.x".}

\item{grid_id_y}{Large size grid identifier used in grid_sf. Default is "GRIDID.y".}

\item{grid_crs}{EPSG code of grid_sf. Default is 3826.}
}
\value{
A data.table including species taxon uuid, grid id, and grid centroid.
}
\description{
Turn coordinates of species occurrence to grid centroid, keeping all columns of species occurrence data
}
\examples{
\dontrun{
# Example usage:
load(system.file("data", "example_occurrence.RData", package = "RangeMap"))
grid_sf <- system.file("data", "grid_1km.RData", package = "RangeMap") \%>\% load
result <- occ2gridCentroid(occ_data = example_occurrence, grid_sf)
}

}
