% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_doc.R
\docType{data}
\name{occ_d}
\alias{occ_d}
\title{Occurrence records for the species Dasypus kappleri}
\format{A data frame with 306 rows and 3 columns.
\describe{
\item{name}{character, species scientific name.}
\item{longitude}{numeric, longitude values.}
\item{latitude}{numeric, latitude values.}
}}
\source{
\url{https://www.gbif.org/}
}
\usage{
occ_d
}
\description{
A dataset containing geographic corrdinates of a South American
armadillo.
}
\examples{
data(occ_d)
head(occ_d)
}
\keyword{datasets}
