% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_doc.R
\docType{data}
\name{occ_f}
\alias{occ_f}
\title{Occurrence records for the species Peltophryne fustiger}
\format{
A data frame with 73 rows and 3 columns.
\describe{
\item{name}{character, species scientific name.}
\item{longitude}{numeric, longitude values.}
\item{latitude}{numeric, latitude values.}
}
}
\source{
\url{https://www.gbif.org/}
}
\usage{
occ_f
}
\description{
A dataset containing geographic coordinates of the Giant Cuban
Toad.
}
\examples{
data("occ_f", package = "rangemap")
head(occ_f)
}
\keyword{datasets}
