% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_doc.R
\docType{data}
\name{occ_p}
\alias{occ_p}
\title{Occurrence records for the species Peltophryne empusa}
\format{
A data frame with 182 rows and 3 columns.
\describe{
\item{name}{character, species scientific name.}
\item{longitude}{numeric, longitude values.}
\item{latitude}{numeric, latitude values.}
}
}
\source{
\url{https://www.gbif.org/}
}
\usage{
occ_p
}
\description{
A dataset containing geographic coordinates of a Caribbean toad.
}
\examples{
data("occ_p", package = "rangemap")
head(occ_p)
}
\keyword{datasets}
