% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_doc.R
\docType{data}
\name{occ_train}
\alias{occ_train}
\title{Occurrence records for the species Amblyomma americanum}
\format{
A data frame with 89 rows and 3 columns.
\describe{
\item{Species}{character, species scientific name.}
\item{Longitude}{numeric, longitude values.}
\item{Latitude}{numeric, latitude values.}
}
}
\source{
\url{https://www.gbif.org/}
}
\usage{
occ_train
}
\description{
A dataset containing geographic coordinates of a North American
tick.
}
\examples{
data("occ_train", package = "rangemap")
head(occ_train)
}
\keyword{datasets}
