% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_init.R
\name{project_init}
\alias{project_init}
\title{Create the project directory structure}
\usage{
project_init(base_dir = getwd())
}
\arguments{
\item{base_dir}{The base directory where the project structure should be created.
The default is the current working directory.}
}
\value{
The function creates the project structure and does not return any value.
}
\description{
This function creates the project folder structure described in the documentation.
}
\examples{
\dontrun{
# To create the project structure in the current working directory:
project_init()

# To create the project structure in a specific directory:
project_init("/path/to/your/directory")

# The directory structure will be created in the specified directory.
}

}
