% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap.R
\docType{package}
\name{rangemap}
\alias{rangemap}
\title{rangemap: An R package for construction of simple species distributional ranges.}
\description{
rangemap presents various tools to create species range maps based on
occurrence data, statistics, and Spatialpolygons. Other tools of
this package can be used to analyze environmental characteristics of
the species ranges and create high quality figures for publication.
}
\section{rangemap functions}{

\code{\link{rangemap_bound}}, \code{\link{rangemap_buff}}, \code{\link{rangemap_enm}},
\code{\link{rangemap_explore}}, \code{\link{rangemap_fig}}, \code{\link{rangemap_hull}},
\code{\link{rangemap_tsa}}, \code{\link{ranges_emaps}}, \code{\link{ranges_espace}}
}

