% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_enm.R
\name{rangemap_enm}
\alias{rangemap_enm}
\title{Species distributional ranges based on ecological niche models}
\usage{
rangemap_enm(occurrences, model, threshold_value, threshold_omission,
    simplify = FALSE, simplify_level = 0, polygons, final_projection,
    save_shp = FALSE, name = "range_enm")
}
\arguments{
\item{occurrences}{a data.frame containing geographic coordinates of species occurrences,
columns must be: Species, Longitude, and Latitude. Geographic coordinates must be in decimal
degrees. \code{occurrences} may not exist but \code{threshold_value} must be defined.}

\item{model}{a RasterLayer object that will be binarized using the \code{threshold_value}
defined by the user or a value calculated based on an omission level (from 0 - 100) defined in
\code{threshold_omission}. If model is projected, this projection must be Geographic (longitude,
latitude). If not projected, the Geographic projection will be assigned for the analysis.}

\item{threshold_value}{(numeric) value used for reclasifying the model. This value will
be the lowest considered as suitable for the species and must be inside the range of values
present in \code{model}. If defined, \code{threshold_omission} will be ignored. If
\code{occurrences} is not defined, this parameter is mandatory.}

\item{threshold_omission}{(numeric) percentage of occurrence records to be excluded from
suitable areas considering their values of suitability in the continuous model (e.g., 0, 5,
or 10). Ignored if \code{threshold_value} is provided.}

\item{simplify}{(logical) if TRUE, polygons of suitable areas will be simplified at a tolerance
defined in \code{simplify_level}. Default = FALSE.}

\item{simplify_level}{(numeric) tolerance at the moment of simplifying polygons created from
the suitable areas derived from the ecological niche model. Lower values will produce polygons
more similar to the original geometry. Default = 0. If simplify is needed, try numbers between
0 and 1 first. Ignored if \code{simplify} = FALSE.}

\item{polygons}{(optional) a SpatialPolygon object to adjust created polygons to these limits.
Projection must be Geographic (longitude, latitude). If not defined, a default, simple world
map will be used.}

\item{final_projection}{(character) string of projection arguments for resulting Spatial objects.
Arguments must be as in the PROJ.4 documentation. See funcion \code{\link[sp]{CRS}} for details.
Default = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0" = WGS84.}

\item{save_shp}{(logical) if TRUE, shapefiles of the species range, occurrences, extent of
occurrence and area of occupancy will be written in the working directory. Default = FALSE.}

\item{name}{(character) valid if \code{save_shp} = TRUE. The name of the shapefile to be exported.
A suffix will be added to \code{name} depending on the object as follows: species extent of
occurrence = "_extent_occ", area of occupancy = "_area_occ", and occurrences = "_unique_records".
Default = "range_enm".}
}
\value{
If \code{occurrences} and \code{threshold_omission} are defined, a  named list containing:
(1) a data.frame with information about the species range, and SpatialPolygon objects of
(2) unique occurrences, (3) species range, (4) extent of occurrence, and (5) area of occurpancy.

If instead of the two parameters before mentioned, \code{threshold_value} is provided, the result
will be a list of two elements: (1) a data.frame with information about the species range, and
(2) a SpatialPolygon object of the species range.
}
\description{
rangemap_enm generates a distributional range for a given species
using a continuous raster layer produced with an ecological niche modeling algorithm.
This function split the model in suitable and unsuitable areas using a user
specified level of omission or a given threshold value. An approach to the species
extent of occurrence (using convex hulls) and the area of occupancy according
to the IUCN criteria are also generated. Shapefiles can be saved in the working
directory if it is needed.
}
\details{
All resultant Spatial objects in the list of results will be projected to the \code{final_projection}.
Areas are calculated in square kilometers using the Azimuthal equal area projection.
}
\examples{
# parameters
sp_mod <- raster::raster(list.files(system.file("extdata", package = "rangemap"),
                                    pattern = "sp_model.tif", full.names = TRUE))
sp_train <- read.csv(list.files(system.file("extdata", package = "rangemap"),
                                pattern = "sp_train.csv", full.names = TRUE))
occ_sp <- data.frame("A_americanum", sp_train)
thres <- 5
save <- TRUE
name <- "test"

enm_range <- rangemap_enm(occurrences = occ_sp, model = sp_mod,  threshold_omission = thres,
                          save_shp = save, name = name)

# see the species range in a figure
extent <- TRUE
occ <- TRUE
legend <- TRUE
north <- TRUE

# creating the species range figure
rangemap_fig(enm_range, add_extent = extent, add_occurrences = occ,
             legend = legend, northarrow = north, northarrow_position = "topleft")
}
