% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_explore.R
\name{rangemap_explore}
\alias{rangemap_explore}
\title{Exploring occurrences before creating range maps}
\usage{
rangemap_explore(occurrences, show_countries = FALSE, graphic_device = FALSE)
}
\arguments{
\item{occurrences}{a data.frame containing geographic coordinates of species occurrences,
columns must be: Species, Longitude, and Latitude. Geographic coordinates must be in
decimal degrees.}

\item{show_countries}{(logical) if TRUE, ISO 3 country codes will label country polygons.}

\item{graphic_device}{(logical) if TRUE, a new graphic device is opened to plot the figure.
Default = FALSE.}
}
\value{
A simple figure of the species occurrences in a geographical context.
}
\description{
rangemap_explore generates simple figures to visualize species occurrence
data in the geographic space.
}
\details{
Base map of countries of the world is a SpatialPolygonDataFrame downloaded from
the Natural Earth database using the \code{\link[rnaturalearth]{ne_countries}} function (scale = 50).
}
\examples{
if(!require(rgbif)){
  install.packages("rgbif")
  library(rgbif)
}

# getting the data from GBIF
species <- name_lookup(query = "Dasypus kappleri",
                       rank="species", return = "data") # information about the species

#occ_count(taxonKey = species$key[17], georeferenced = TRUE) # testing if keys return records

key <- species$key[17] # using species key that return information

occ <- occ_search(taxonKey = key, return = "data") # using the taxon key

# keeping only georeferenced records
occ_g <- occ[!is.na(occ$decimalLatitude) & !is.na(occ$decimalLongitude),
             c("name", "decimalLongitude", "decimalLatitude")]

# simple figure of the species occurrence data
explore_map <- rangemap_explore(occurrences = occ_g, show_countries = TRUE)

#dev.off() # for returning to default par settings
}
