% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_fig.R
\name{rangemap_fig}
\alias{rangemap_fig}
\title{Figures of species range maps}
\usage{
rangemap_fig(range, polygons, add_extent = FALSE, add_occurrences = FALSE,
  grid = FALSE, sides = "bottomleft", northarrow = FALSE,
  northarrow_position = "topright", scalebar = FALSE,
  scalebar_position = "bottomleft")
}
\arguments{
\item{range}{an object produced with any of the following functions:
\code{\link{rangemap_buff}}, \code{\link{rangemap_bound}}, \code{\link{rangemap_hull}},
\code{\link{rangemap_enm}}, and \code{\link{rangemap_tsa}}.}

\item{polygons}{a SpatialPolygon object to be used as base map for plotting the species range.
If not provided, a simplified world map will be used.}

\item{add_extent}{(logical) if TRUE the extent of occurrence of the species will be added to
the figure.}

\item{add_occurrences}{(logical) if TRUE the species occurrence records will be added to
the figure.}

\item{grid}{(logical) if TRUE labels and ticks of coordinates will be inserted in sides.}

\item{sides}{(character) sides in which the labels will be placed in the figure. Options
for this are the same than for other position character options indicators.}

\item{northarrow}{(logical) if TRUE, a simple north arrow will be placed in northarrow_position.}

\item{northarrow_position}{(character) site in the figure where the north arrow will be placed.}

\item{scalebar}{(logical) if TRUE a simple scale bar will be inserted in scalebar_position.}

\item{scalebar_position}{(character) place for the scale bar insertion.}
}
\value{
A figure of the species distributional range in a geographical context, with the
map components defined by the user.
}
\description{
rangemap_fig generates customizable figures of species range maps
using the objects produced by other function of this package.
}
\details{
Position of distinct elements depend on the spatial configuration of the species range.
Therefore, their positiuon may need to be changed if the elements are needed. Position options are:
"bottomright", "bottomleft", "topleft", and "topright".
}
\examples{

if(!require(rgbif)){
  install.packages("rgbif")
  library(rgbif)
}

# getting the data from GBIF
occ <- occ_search(taxonKey = 2440788, return = "data")

# keeping only georeferenced records
occ_g <- occ[!is.na(occ$decimalLatitude) & !is.na(occ$decimalLongitude),
             c("name", "decimalLongitude", "decimalLatitude")]

level <- 0
dissolve <- FALSE
save <- FALSE
countries <- c("PER", "BRA", "COL", "VEN", "ECU", "GUF", "GUY", "SUR", "BOL")

range <- rangemap::rangemap_bound(occurrences = occ_g, country_code = countries, boundary_level = level,
                                dissolve = FALSE, save_shp = FALSE)

extent <- TRUE
occ <- TRUE
grid <- TRUE
sides <- "bottomleft"

range_map <- rangemap_fig(range, add_extent = extent, add_occurrences = occ,
                          grid = grid, sides = sides)
}
