% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_hull.R
\name{rangemap_hull}
\alias{rangemap_hull}
\title{Species distributional ranges based on distinct hull polygons}
\usage{
rangemap_hull(occurrences, hull_type = "concave", distance = 0.5,
  split_distance = 2, export = FALSE)
}
\arguments{
\item{occurrences}{a data.frame containing species occurrences, columns must be:
Species, Longitude, and Latitude.}

\item{hull_type}{(character) type of hull polygons to be created. Available options are: "convex",
"concave", and "alpha" hulls.}

\item{distance}{(numeric) distance in decimal degrees to be used for creating a buffer area
around the hull polygons, default = 0.5.}

\item{split_distance}{(numeric) distance in decimal degrees that will limit connectivity among
hull polygons created with chunks of points separated by long distances, , default = 2.}

\item{export}{(logical) if TRUE a shapefile of the species range will be written in the working
directory, appart of the returned object.}
}
\value{
A named list containing a data.frame with information about the species range, a
SpatialPolygon object of the species range in Geographic projection, and the same SpatialPolygon
object projected to the Azimuthal equal area projection.
}
\description{
rangemap_hull generates a species range polygon for a given species
by considering all the polygons of political entities in which the species has
been detected.
}
\examples{
if(!require(rgbif)){
install.packages("rgbif")
library(rgbif)
}

# getting the data from GBIF
occ <- occ_search(taxonKey = 5219426, return = "data", limit = 1000)

# keeping only georeferenced records
occ_g <- occ[!is.na(occ$decimalLatitude) & !is.na(occ$decimalLongitude),]

disolve <- TRUE
hull <- "convex"
split <- 4

hull_range <- rangemap_hull(occurrences = occ_g, hull_type = hull, distance = dist, split_distance = split)
}
