% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_tsa.R
\name{rangemap_tsa}
\alias{rangemap_tsa}
\title{Species distributional ranges based on a trend surface analysis}
\usage{
rangemap_tsa(occurrences, region_of_interest, threshold = 0,
  save_shp = FALSE, name)
}
\arguments{
\item{occurrences}{a data.frame containing species occurrences, columns must be:
Species, Longitude, and Latitude.}

\item{region_of_interest}{a SpatialPolygon object on which the trend surface analysis
will be performed.}

\item{threshold}{(numeric) percentage of occurrence records to be excluded when deciding
the minimum surface value to be considered part of the species range, default = 0.}

\item{save_shp}{(logical) if TRUE shapefiles of the species range, extent of occurrence and area of
occupancy will be written in the working directory.}

\item{name}{(character) valid if save_shp TRUE. The name of the shapefile to be exported.}

\item{resolution}{(numeric) resolution in km in which the resultant surface will be created,
default = 5.}
}
\value{
A named list containing a data.frame with information about the species range, a
SpatialPolygon object of the species range in Geographic projection, and the same SpatialPolygon
object projected to the Azimuthal equal area projection.
}
\description{
rangemap_tsa generates species range polygons for a given species
using .
}
\details{
Trend surface analysis Is a method based on low-order polynomials of spatial coordinates
for estimating a regular grid of points from scattered observations. This method assumes that all
cells not occupied by occurrences are absences; hence its use depends on the quality of data and
the certainty of having or not a complete sampling of the regiong_of_interest.
}
