% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_envcomp.R
\name{ranges_envcomp}
\alias{ranges_envcomp}
\title{Comparison of species ranges in environmental space}
\usage{
ranges_envcomp(occurrences, variables, ranges, export = FALSE,
  format = "png", ...)
}
\arguments{
\item{occurrences}{a data.frame containing species occurrences, columns must be:
Species, Longitude, and Latitude.}

\item{variables}{a RasterStack object of environmental variables that will be used for
performing the trend surface analysis.}

\item{ranges}{a list of objects produced with any of the following functions:
\code{\link{rangemap_buff}}, \code{\link{rangemap_bound}}, \code{\link{rangemap_hull}},
\code{\link{rangemap_enm}}, and \code{\link{rangemap_tsa}}.}

\item{export}{(logical) if TRUE a figure in format = format will be written in the working
directory, appart of the returned object.}

\item{format}{(character) format of the figure that will be written in the working directory
if export = TRUE.}

\item{...}{other arguments from function \code{\link[Base]{plot}}.}
}
\value{
A figure showing, in the environmental space, the species ranges generated with any
of the functions: \code{\link{rangemap_buff}}, \code{\link{rangemap_bound}},
\code{\link{rangemap_hull}}, \code{\link{rangemap_enm}}, and \code{\link{rangemap_tsa}}.
}
\description{
ranges_envcomp generates a two dimensional comparison of a species'
ranges created using distinct algortihms, to visualize implications of selecting
one of them if environmental conditions are considered.
}
\details{
Trend surface analysis Is a method based on low-order polynomials of spatial coordinates
for estimating a regular grid of points from scattered observations.
}
\examples{
# comming
}
