% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_espace.R
\name{save_rgs_espace}
\alias{save_rgs_espace}
\title{Helper function to save plotly figures.}
\usage{
save_rgs_espace(p, name = "ranges_espace", width = 1000, height = 800)
}
\arguments{
\item{p}{object containing the plotly plot.}

\item{name}{(character) name of the figure to be saved, default = "ranges_espace".}

\item{width}{(numeric) width of the figure in pixels, default = 4150.}

\item{height}{(numeric) height of the figure in pixels, default = 3320.}
}
\description{
Helper function to save plotly figures.
}
