% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_fig.R
\name{scalebarf}
\alias{scalebarf}
\title{Helper function to plot a scale bar.}
\usage{
scalebarf(loc, length, unit = "km", division.cex = 0.8, ...)
}
\arguments{
\item{loc}{(numeric) vector (x, y) of coordinates where the scale bar will be plotted.}

\item{length}{(numeric) length of the scale bar. Try to use numbers divisive to 2.}

\item{unit}{(character) units in which the scale will be represented. Default "km"}

\item{division.cex}{(numeric) cex of the scale components.}

\item{...}{other graphical arguments.}
}
\description{
Helper function to plot a scale bar.
}
\details{
this function is a modification of the "scalebar" function developed by T
animura et al. (2007) (see \url{http://hdl.handle.net/10.18637/jss.v019.c01}).
}
