% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simple_wmap}
\alias{simple_wmap}
\title{Get a simplified SpatialPolygons of the world}
\usage{
simple_wmap(which = "simplest")
}
\arguments{
\item{which}{(charatcer) name of type of SpatialPolygons to be obtained. Options
are: "simple" and "simplest"; default = "simplest".}
}
\description{
Get a simplified SpatialPolygons of the world
}
\examples{
map <- simple_wmap()
}
