% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_doc.R
\name{sp_model}
\alias{sp_model}
\title{An ecological niche model created with Maxent}
\format{
A RasterLayer with 150 rows, 249 columns, and 37350 cells:
\describe{
\item{Suitability}{suitability, in probability values.}
}
}
\source{
\url{https://kuscholarworks.ku.edu/handle/1808/26376}
}
\value{
No return value, used with function \code{\link[raster]{raster}} to
bring an example of ecological niche modeling output.
}
\description{
A RasterLayer containing an ecological niche model for the a tick
(\emph{Amblyomma americanum}).
}
\examples{
model <- raster::raster(system.file("extdata", "sp_model.tif",
                                    package = "rangemap"))

raster::plot(model)
}
