% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{sp_range}
\alias{sp_range}
\alias{sp_range-class}
\alias{sp_range_iucn-class}
\alias{sp_range_iucnextra-class}
\alias{sp_range_iucn}
\alias{sp_range_iucnextra}
\title{An S4 class to organize data and results of sp_range* objects}
\description{
A list of classes (some of them inherited) to contain information
derived from analyses using \code{\link{rangemap}}. Available classes are:
sp_range, sp_range_iucn, and sp_range_iucnextra.
}
\section{Slots}{

\describe{
\item{\code{Summary}}{a data.frame.}

\item{\code{Species_unique_records}}{("numeric"), not in sp_range.}

\item{\code{Species_range}}{a SpatialPolygonsDataFrame.}

\item{\code{Extent_of_occurrence}}{a SpatialPolygonsDataFrame, not in sp_range.}

\item{\code{Area_of_occupancy}}{a SpatialPolygonsDataFrame, not in sp_range.}

\item{\code{Trend_surface_model}}{a RasterLayer, not in sp_range.}
}}

\examples{
showClass("sp_range")
showClass("sp_range_iucn")
}
