% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_doc.R
\docType{data}
\name{spdf_range}
\alias{spdf_range}
\title{Example SpatialPolygonsDataFrame of a species range}
\format{
SpatialPolygonsDataFrame with 1 feature.
\describe{
\item{features}{SpatialPolygons, 1.}
\item{data.frame}{ID = 1.}
}
}
\usage{
spdf_range
}
\description{
A SpatialPolygonsDataFrame representing the distribution of a
species from North America.
}
\examples{
data("spdf_range", package = "rangemap")
spdf_range
}
\keyword{datasets}
