% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary,sp_range-method}
\alias{summary,sp_range_iucn-method}
\alias{summary,sp_range_iucnextra-method}
\title{Summary of attributes and results}
\usage{
\S4method{summary}{sp_range}(object)

\S4method{summary}{sp_range_iucn}(object)

\S4method{summary}{sp_range_iucnextra}(object)
}
\arguments{
\item{object}{object of class sp_range*.}
}
\value{
A written summary.
}
\description{
Summary of attributes and results
}
