% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{tw_main}
\alias{tw_main}
\title{Data of layer of Taiwan main island}
\format{
A sf polygon.
}
\source{
Polygon of Taiwan.
}
\usage{
tw_main
}
\description{
This data contains polygon of Taiwan main island
}
\keyword{datasets}
