% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_attribute_layers.R
\name{generate_attribute_layers}
\alias{generate_attribute_layers}
\title{Generate raster layers of individual attributes.}
\usage{
generate_attribute_layers(
  raster_path,
  attributes_path,
  attribute_names,
  AOI,
  output_directory,
  n_cores = NULL,
  tile_size_adjustment = 1
)
}
\arguments{
\item{raster_path}{Path to RangeMap raster file for a single year}

\item{attributes_path}{Path to attributes table file. It is preferable to use the RangeMap_Attributes.csv so full field names are preserved, but a tif.vat.dbf file associated with one year's raster may be used}

\item{attribute_names}{Names of attributes desired for generating raster layers. These must match names in the attributes table file}

\item{AOI}{Area of interest. Spatial area for generating raster layers. This can be a polygon or raster loaded into environment (as terra of sf object), or a file path to a .shp or .tif}

\item{output_directory}{Full directory path for the output rasters (not a file path). Does not need to already be created}

\item{n_cores}{If desired, set the number of cores to use (defaults to 40\% of total cores)}

\item{tile_size_adjustment}{If desired, adjust tile sizing. Set this to less than 1 to create smaller tiles less prone to failure in processing}
}
\value{
Raster file of all desired attributes
}
\description{
Generate raster layers of individual attributes.
}
