% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_attribute_names.R
\name{load_attribute_names}
\alias{load_attribute_names}
\title{Load an attributes file (.csv or .dbf) and get attribute names}
\usage{
load_attribute_names(attributes_path)
}
\arguments{
\item{attributes_path}{Path to attributes file. Preferable to use the RangeMap_Attribute.csv so full field names are preserved, but the .tif.vat.dbf associated with one year's raster also works.}
}
\value{
Names of all numeric attributes that can be generated
}
\description{
Load an attributes file (.csv or .dbf) and get attribute names
}
