% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap.R
\docType{package}
\name{rangemap}
\alias{rangemap}
\alias{rangemap-package}
\title{rangemap: A package for construction of species distributional range maps.}
\description{
rangemap presents various tools to create species range maps based on
occurrence data, statistics, and distinct shapefiles. Other tools of
this package can be used to analyze environmental characteristics of
the species ranges and create high quality figures for publication.
}
\section{rangemap functions}{

\code{\link{rangemap_bound}}, \code{\link{rangemap_buff}}, \code{\link{rangemap_enm}},
\code{\link{rangemap_explore}}, \code{\link{rangemap_fig}}, \code{\link{rangemap_hull}},
\code{\link{rangemap_tsa}}, \code{\link{ranges_emaps}}, \code{\link{ranges_espace}}
}

