% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_buff.R
\name{rangemap_buff}
\alias{rangemap_buff}
\title{Species distributional ranges based on buffered occurrences}
\usage{
rangemap_buff(occurrences, distance = 1, polygons, export = FALSE)
}
\arguments{
\item{occurrences}{a data.frame containing species occurrences, columns must be:
Species, Longitude, and Latitude.}

\item{distance}{(numeric) distance in decimal degrees to be used for creating the buffer areas
around occurrences, default = 1.}

\item{polygons}{(optional) a SpatialPolygon object that will be clipped with the buffer areas
to create species ranges based on actual limits. If not defined, a default world map will be
used.}

\item{export}{(logical) if TRUE a shapefile of the species range will be written in the working
directory, appart of the returned object.}
}
\value{
A named list containing a data.frame with information about the species range, a
SpatialPolygon object of the species range in Geographic projection, and the same SpatialPolygon
object projected to the Azimuthal equal area projection.
}
\description{
rangemap_buff generates a species range polygon for a given species
by buffering provided occurrences using a defined distance. Shape files can be saved
in the working directory if it is needed.
}
\examples{
if(!require(rgbif)){
install.packages("rgbif")
library(rgbif)
}

# getting the data from GBIF
occ <- occ_search(taxonKey = 2422480, return = "data")

# keeping only georeferenced records
occ_g <- pe_dat[!is.na(pe_dat$decimalLatitude) & !is.na(pe_dat$decimalLongitude),]

# buffer distance
dist <- 0.5

buff_range <- rangemap_buff(occurrences = occ_g, distance = dist)
}
