% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_buff.R
\name{rangemap_buff}
\alias{rangemap_buff}
\title{Species distributional ranges based on buffered occurrences}
\usage{
rangemap_buff(occurrences, buffer_distance = 1e+05, polygons, final_projection,
    save_shp = FALSE, name = "range_buffer", overwrite = FALSE)
}
\arguments{
\item{occurrences}{a data.frame containing geographic coordinates of species occurrences,
columns must be: Species, Longitude, and Latitude. Geographic coordinates must be in decimal
degrees.}

\item{buffer_distance}{(numeric) distance, in meters, to be used for creating the buffer areas
around occurrences, default = 100000.}

\item{polygons}{(optional) a SpatialPolygon object to clip buffer areas and adjust the species
range and other polygons to these limits. Projection must be Geographic (longitude, latitude).
If not defined, a default, simple world map will be used.}

\item{final_projection}{(character) string of projection arguments for resulting Spatial objects.
Arguments must be as in the PROJ.4 documentation. See \code{\link[sp]{CRS-class}} for details.
Default = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0" = WGS84.}

\item{save_shp}{(logical) if TRUE, shapefiles of the species range, occurrences, extent of
occurrence and area of occupancy will be written in the working directory. Default = FALSE.}

\item{name}{(character) valid if \code{save_shp} = TRUE. The name of the shapefile to be
exported. A suffix will be added to \code{name} depending on the object as follows: species
extent of occurrence = "_extent_occ", area of occupancy = "_area_occ", and occurrences =
"_unique_records". Default = "range_buffer".}
}
\value{
A named list containing: (1) a data.frame with information about the species range,
and SpatialPolygon objects of (2) unique occurrences, (3) species range, (4) extent of
occurrence, and (5) area of occurpancy.
}
\description{
rangemap_buff generates a distributional range for a given species by
buffering provided occurrences using a defined distance. An approach to the species extent
of occurrence (using convex hulls) and the area of occupancy according to the IUCN criteria
are also generated. Shapefiles can be saved in the working directory if it is needed.
}
\details{
All resultant Spatial objects in the list of results will be projected to the \code{final_projection}.
Areas are calculated in square kilometers using the Azimuthal equal area projection.
}
\examples{
suppressWarnings({if(!require(spocc)){
  install.packages("spocc")
  library(spocc)
}})

# getting the data from GBIF
occs <- occ(query = "Peltophryne empusa", from = "gbif",
            limit = 1000)$gbif$data[[1]]

# keeping only georeferenced records
occ_g <- occs[!is.na(occs$latitude) & !is.na(occs$longitude),
              c("name", "longitude", "latitude")]

# buffer distance
dist <- 100000
save <- TRUE
name <- "test"

buff_range <- rangemap_buff(occurrences = occ_g, buffer_distance = dist,
                            save_shp = save, name = name, overwrite = TRUE)

# see the species range in a figure
extent <- TRUE
occ <- TRUE
legend <- TRUE
north <- TRUE

# creating the species range figure
rangemap_fig(buff_range, add_extent = extent, add_occurrences = occ,
             legend = legend, northarrow = north, legend_position = "bottomleft")
}
