% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_enm.R
\name{rangemap_enm}
\alias{rangemap_enm}
\title{Species distributional ranges based on ecological niche models}
\usage{
rangemap_enm(occurrences, model, threshold_value, threshold,
  simplify_level = 0, polygons, save_shp = FALSE, name)
}
\arguments{
\item{occurrences}{a data.frame containing species occurrences, columns must be:
Species, Longitude, and Latitude.}

\item{model}{a RasterLayer object that will be binarized using the threshold value defined
by the user or a value calculated based on a value (from 0 - 100) defined in threshold.}

\item{threshold_value}{(numeric) decimal value used for reclasifying the model. This value will
be the lowest considered as suitable for the species.}

\item{threshold}{(numeric) percentage of occurrence records to be excluded from suitable areas
considering their values of suitability in the continuous model (e.g., 0, 5, or 10).}

\item{simplify_level}{(numeric) tolerance at the moment of simplifying polygons created from
the suitable areas derived from the ecological niche model. Lower values will produce polygons
more similar to the original geometry. Default = 0. If simplify is needed, try numbers between
0 and 1 first.}

\item{polygons}{(optional) a SpatialPolygon object that will be clipped with the buffer areas
to create species ranges based on actual limits. Projection must be Geographic (longitude, latitude).
If not defined, a default, simple world map will be used.}

\item{save_shp}{(logical) if TRUE shapefiles of the species range, extent of occurrence and area of
occupancy will be written in the working directory, default = FALSE.}

\item{name}{(character) valid if save_shp TRUE. The name of the shapefile to be exported.}
}
\value{
A named list containing a data.frame with information about the species range, a
SpatialPolygon object of the species range in Geographic projection, and the same SpatialPolygon
object projected to the Azimuthal equal area projection.
}
\description{
rangemap_enm generates species range polygons for a given species
using a continuous raster layer produced with an ecological niche modeling tool.
This function split the model in suitable and unsuitable habitats using a user
specified level of error or a given threshold value. An approach to the species
extent of occurrence (using convex hulls) and the area of occupancy according
to the IUCN criteria are also generated. Shape files can be saved in the working
directory if it is needed.
}
\details{
If threshold_value is provided, argument threshold is ignored.
}
\examples{
if(!require(devtools)){
install.packages("devtools")
library(devtools)
}

if(!require(kuenm)){
install_github("marlonecobos/kuenm")
library(kuenm)
}

# parameters
data(sp_mod)
data(sp_train)
occ_sp <- data.frame("A_americanum", sp_train)
thres <- 5
save <- TRUE
name <- "test"

enm_range <- rangemap_enm(occurrences = occ_sp, model = sp_mod,  threshold = thres,
                          save_shp = save, name = name)
}
