% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_explore.R
\name{rangemap_explore}
\alias{rangemap_explore}
\title{Figures for exploring occurrences on maps before creating range maps}
\usage{
rangemap_explore(occurrences, polygons, graphic_device = FALSE)
}
\arguments{
\item{occurrences}{a data.frame containing species occurrences, columns must be:
Species, Longitude, and Latitude.}

\item{polygons}{a SpatialPolygon object to be used as base map for plotting the species range.
If not provided, a simplified world map will be used.}
}
\value{
A simple figure of the species occurrences in a geographical context.
}
\description{
rangemap_explore generates simple figures to visualize species occurrence
data in the geographic space before using other function sof this package.
}
\examples{

if(!require(rgbif)){
  install.packages("rgbif")
  library(rgbif)
}

# getting the data from GBIF
species <- name_lookup(query = "Dasypus kappleri",
                       rank="species", return = "data") # information about the species

occ_count(taxonKey = species$key[14], georeferenced = TRUE) # testing if keys return records

key <- species$key[14] # using species key that return information

occ <- occ_search(taxonKey = key, return = "data") # using the taxon key

# keeping only georeferenced records
occ_g <- occ[!is.na(occ$decimalLatitude) & !is.na(occ$decimalLongitude),
             c("name", "decimalLongitude", "decimalLatitude")]

# simple figure of the species occurrence data
explore_map <- rangemap_explore(occurrences = occ_g)

dev.off() # for returning to default par settings
}
