% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemap_tsa.R
\name{rangemap_tsa}
\alias{rangemap_tsa}
\title{Species distributional ranges based on a trend surface analysis}
\usage{
rangemap_tsa(occurrences, variables, threshold = 5, export = FALSE)
}
\arguments{
\item{occurrences}{a data.frame containing species occurrences, columns must be:
Species, Longitude, and Latitude.}

\item{variables}{a RasterStack object of environmental variables that will be used for
performing the trend surface analysis.}

\item{threshold}{(numeric) percentage of occurrence records to be excluded when deciding
the minimum surface value to be considered part of the species range, default = 5.}

\item{export}{(logical) if TRUE a shapefile of the species range will be written in the working
directory, appart of the returned object.}
}
\value{
A named list containing a data.frame with information about the species range, a
SpatialPolygon object of the species range in Geographic projection, and the same SpatialPolygon
object projected to the Azimuthal equal area projection.
}
\description{
rangemap_tsa generates species range polygons for a given species
using .
}
\details{
Trend surface analysis Is a method based on low-order polynomials of spatial coordinates
for estimating a regular grid of points from scattered observations.
}
