% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simple_wmap}
\alias{simple_wmap}
\title{Get a simplified SpatialPolygonsDataFrame of the world}
\usage{
simple_wmap(which = "simplest", regions = ".")
}
\arguments{
\item{which}{(character) name of type of SpatialPolygons to be obtained. Options
are: "simple" and "simplest"; default = "simplest".}

\item{regions}{(character) name of the country (or region if \code{which} =
"simple") for which to obtain the SpatialPolygonsDataFrame.}
}
\value{
A simplified SpatialPolygonsDataFrame of the world in WGS84 projection.
}
\description{
Get a simplified SpatialPolygonsDataFrame of the world
}
\examples{
map <- simple_wmap()
}
